Update README.md
As requested by Louise the following analysis it will demonstrate insights and statistics results based on the Kickstarter data provided in order to help Louise develop the new Theater Campaign.

 
 
As in our analysis perception the Theater category was more successful than other category’s, reaffirming that Louise interest on the Theater category has potential. Furthermore, we can identify what would be recommended in order to obtained success on the project.

 
 
Is strongly recommended that Louise start her Theater Play Campaigned in the months of May, June, and July. Looking at the lower success rates falling in January, February, August, September, and October, with the lowest rates in December. It should be considered starting a theatre campaign in Summer as that suggests the highest success rate over other times of the year.



 

For the Great Britain's theater market, especially musicals. After close detail analyses of the amount estimated budget of £4,000, we recommend request a lower amount for budget possibly from £2000.00 to £3000.00, considering half of the campaign goals are less than £2,000.
Also, in the Great Britain’s the category Theater, as in US, is the most successful. Indicating that the Theater market is a good investment choice.







Observations:
On US campaigns the success average goal was close to pledge amount as or IQR statistics number showed, however on the failed campaigns the discrepancies are uncommonly high.
Consider check the Goal and Pledge for failed campaigns once we had found considerable amount differences.
Theater US
 	 Successful 	 Failed 
IQR of Goal	 $   3,500.00 	 $   8,000.00 
IQR of Pledged	 $   3,982.49 	 $      491.75 
		
Theater Spaces/Musical US
 	 Successful 	 Failed 
IQR of Goal	 $ 17,648.89 	 $ 45,000.00 
IQR of Pledged	 $ 17,648.89 	 $   1,504.00